Fix windows build - cast int port to uint16_t

+++ source/cpp/source/TestCentre.cpp
@@ -37,7 +37,7 @@ public:
             auto portFromArgs = getPort ();
             if (! portFromArgs)
                 return;
-            port = *portFromArgs;
+            port = static_cast<uint16_t>(*portFromArgs);
         }
 
         addCommandHandler (_defaultCommandHandler);
